eng_soft/tp/gpt_api/requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.2.0
autogui==0.1.8
Babel==2.14.0
backcall==0.2.0
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
boto3==1.28.64
botocore==1.31.64
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1666764671472/work
cachetools==5.2.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.1.0
cramjam==2.8.3
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography-split_1691444160254/work
cycler==0.11.0
dataclasses-json==0.5.14
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
effdet==0.4.1
emoji==2.8.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.103.0
fastjsonschema==2.19.1
fastparquet==2024.2.0
filelock==3.12.2
filetype==1.2.0
Flask==3.0.3
Flask-Cors==4.0.0
flatbuffers==23.1.4
fonttools==4.41.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.1
h11==0.14.0
h5py==3.7.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1688754491823/work
iopath==0.1.10
ipykernel==6.20.2
ipython==8.8.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
keras==2.11.0
keyboard==0.13.5
kiwisolver==1.4.4
langchain==0.0.263
langsmith==0.0.22
layoutparser==0.3.4
Levenshtein==0.23.0
libclang==15.0.6.1
lxml==4.9.3
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.2
mlxtend==0.22.0
MouseInfo==0.1.3
mpmath==1.2.1
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
notebook==7.1.2
notebook_shim==0.2.4
numexpr==2.8.5
numpy==1.25.1
omegaconf==2.3.0
onnxruntime==1.15.1
openai==1.3.4
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.76
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.0
pandas==1.5.2
pandas-stubs==2.0.2.230605
pandocfilters==1.5.1
parso==0.8.3
pdf2image==1.16.3
pdfminer==20191125
pdfminer.six==20221105
pdfplumber==0.10.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare==0.7.5
pillow==10.2.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1690813113769/work
portalocker==2.7.0
prometheus_client==0.20.0
prompt-toolkit==3.0.36
protobuf==4.24.2
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1681775027942/work
psycopg2==2.9.7
psycopg2-binary==2.9.7
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval==0.2.2
pyarrow==16.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.54
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.12
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.14.0
PyMsgBox==1.0.9
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1685514481738/work
pyparsing==3.0.9
pypdf==3.15.2
pypdfium2==4.18.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-Levenshtein==0.23.0
python-magic==0.4.27
python-multipart==0.0.6
python3-xlib==0.15
pythonnet==3.0.3
pytweening==1.0.7
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.0.0
rapidfuzz==3.4.0
referencing==0.34.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.7.0
safetensors==0.3.3
scikit-learn==1.3.2
scikit-surprise==1.1.3
scipy==1.11.1
seaborn==0.13.0
selenium==4.17.2
Send2Trash==1.8.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.27.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
terminado==0.18.1
threadpoolctl==3.1.0
tiktoken==0.4.0
timm==0.9.5
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0+cpu
torch-tb-profiler==0.4.1
torchaudio==2.0.1+cpu
torchvision==0.15.1+cpu
tornado==6.2
tqdm==4.65.0
traitlets==5.8.1
transformers==4.32.0
trio==0.24.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
types-pytz==2023.3.0.0
typing-inspect==0.9.0
typing_extensions==4.11.0
unstructured==0.10.8
unstructured-inference==0.5.17
uri-template==1.3.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1678635778344/work
uvicorn==0.23.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1689374466814/work
